+++ backend/windmill-worker/src/lib.rs
@@ -36,6 +36,7 @@ pub mod job_logger;
 pub mod job_logger_ee;
 mod job_logger_oss;
 mod js_eval;
+mod memory_common;
 #[cfg(feature = "private")]
 pub mod memory_ee;
 mod memory_oss;
